refactor(dsl): upgrage patcher and refactor patches to use new apis (#66)
* refactor: build

* refactor: build to dsl

* refactor: more files migrated

* refactor: more files migrated

* refactor: complete migration to dsl

* fix: Bump gradle version

* refactor: format code

* fix: remove kls database file

* fix: add kls file to gitignore

* fix: upstream build

* build: upgrade wrapper

* fix

* remove unnessary tasks

* fix: gen bundle script

* refactor: cleanup workflows

* build: bump deps
