+++ .github/workflows/release.yml
@@ -5,7 +5,7 @@ on:
   workflow_call:
   push:
     branches:
-      # - main
+      - main
       - dev
 
 jobs:
@@ -46,6 +46,13 @@ jobs:
           fingerprint: ${{ env.GPG_FINGERPRINT }}
 
       - name: Release
+        id: semantic_release 
         env:
           GITHUB_TOKEN: ${{ secrets.REPOSITORY_PUSH_ACCESS }}
         run: npm exec semantic-release
+        
+      - name: Call Update Bundle Workflow
+        if: github.ref == 'refs/heads/main'  # Check if the branch is main
+        uses: ./.github/workflows/update-bundle.yml  # Path to the update-bundle workflow
+        with:
+          version: ${{ steps.semantic_release.outputs.version }}  # Pass the version output
