--- .github/workflows/release_main.yml
@@ -1,16 +1,0 @@
-name: publish release_main
-
-on:
-  workflow_dispatch:
-  push:
-    branches:
-      - main
-jobs:
-  build:
-    uses: indrastorms/dropped-patches/.github/workflows/release.yml@main
-    secrets: inherit
-
-  update_bundle:
-    needs: [build]
-    uses: indrastorms/dropped-patches/.github/workflows/update_bundle.yml@main
-    secrets: inherit
