--- api/dropped-patches.api
@@ -1,51 +1,0 @@
-public final class dropped/patches/fxfileexplorer/unlockplus/UnlockPlusPatch : app/revanced/patcher/patch/BytecodePatch {
-	public static final field INSTANCE Ldropped/patches/fxfileexplorer/unlockplus/UnlockPlusPatch;
-	public fun execute (Lapp/revanced/patcher/data/BytecodeContext;)V
-	public synthetic fun execute (Lapp/revanced/patcher/data/Context;)V
-}
-
-public final class dropped/patches/fxfileexplorer/unlockplus/fingerprints/IsPlusUnlockedFingerprint : app/revanced/patcher/fingerprint/MethodFingerprint {
-	public static final field INSTANCE Ldropped/patches/fxfileexplorer/unlockplus/fingerprints/IsPlusUnlockedFingerprint;
-}
-
-public final class dropped/patches/nova/prime/fingerprints/UpdatePrimeStatusFingerprint : app/revanced/patcher/fingerprint/MethodFingerprint {
-	public static final field INSTANCE Ldropped/patches/nova/prime/fingerprints/UpdatePrimeStatusFingerprint;
-}
-
-public final class dropped/patches/nova/prime/patch/UnlockPrimePatch : app/revanced/patcher/patch/BytecodePatch {
-	public static final field INSTANCE Ldropped/patches/nova/prime/patch/UnlockPrimePatch;
-	public fun execute (Lapp/revanced/patcher/data/BytecodeContext;)V
-	public synthetic fun execute (Lapp/revanced/patcher/data/Context;)V
-}
-
-public abstract class dropped/patches/shared/misc/hex/BaseHexPatch : app/revanced/patcher/patch/RawResourcePatch {
-	public fun <init> ()V
-	public synthetic fun execute (Lapp/revanced/patcher/data/Context;)V
-	public fun execute (Lapp/revanced/patcher/data/ResourceContext;)V
-	public abstract fun getReplacements ()Ljava/util/List;
-}
-
-public final class dropped/patches/shared/misc/hex/BaseHexPatch$Replacement {
-	public static final field Companion Ldropped/patches/shared/misc/hex/BaseHexPatch$Replacement$Companion;
-	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
-	public final fun replacePattern ([B)V
-}
-
-public final class dropped/patches/shared/misc/hex/BaseHexPatch$Replacement$Companion {
-}
-
-public final class dropped/patches/spotify/premium/UnlockPremiumPatch : dropped/patches/shared/misc/hex/BaseHexPatch {
-	public fun <init> ()V
-	public fun getReplacements ()Ljava/util/List;
-}
-
-public final class dropped/patches/tasker/trial/unlock/fingerprints/CheckLicenseFingerprint : app/revanced/patcher/fingerprint/MethodFingerprint {
-	public static final field INSTANCE Ldropped/patches/tasker/trial/unlock/fingerprints/CheckLicenseFingerprint;
-}
-
-public final class dropped/patches/tasker/trial/unlock/patch/UnlockLicensePatch : app/revanced/patcher/patch/BytecodePatch {
-	public static final field INSTANCE Ldropped/patches/tasker/trial/unlock/patch/UnlockLicensePatch;
-	public fun execute (Lapp/revanced/patcher/data/BytecodeContext;)V
-	public synthetic fun execute (Lapp/revanced/patcher/data/Context;)V
-}
-
