+++ gradle/libs.versions.toml
@@ -1,15 +1,4 @@
 [versions]
-revanced-patcher = "19.3.1"
-revanced-patches = "4.8.0-dev.8"
+revanced-patcher = "20.0.2"
+# revanced-patches = "4.8.0-dev.8"
 smali = "3.0.5"
-binary-compatibility-validator = "0.14.0"
-kotlin = "1.9.22"
-
-[libraries]
-revanced-patcher = { module = "app.revanced:revanced-patcher", version.ref = "revanced-patcher" }
-revanced-patches = { module = "app.revanced:revanced-patches", version.ref = "revanced-patches" }
-smali = { module = "com.android.tools.smali:smali", version.ref = "smali" }
-
-[plugins]
-binary-compatibility-validator = { id = "org.jetbrains.kotlinx.binary-compatibility-validator", version.ref = "binary-compatibility-validator" }
-kotlin = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
