+++ package.json
@@ -3,7 +3,7 @@
     "@saithodev/semantic-release-backmerge": "^4.0.1",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
-    "gradle-semantic-release-plugin": "^1.9.1",
-    "semantic-release": "^23.0.7"
+    "gradle-semantic-release-plugin": "^1.9.2",
+    "semantic-release": "^24.0.0"
   }
 }
