+++ patches/api/patches.api
@@ -1,0 +1,29 @@
+public final class dropped/patches/fxfileexplorer/unlockplus/UnlockPlusPatchKt {
+	public static final fun getUnlockPlusPatch ()Lapp/revanced/patcher/patch/BytecodePatch;
+}
+
+public final class dropped/patches/nova/prime/UnlockPrimePatchKt {
+	public static final fun getUnlockPrimePatch ()Lapp/revanced/patcher/patch/BytecodePatch;
+}
+
+public final class dropped/patches/shared/misc/hex/HexPatchKt {
+	public static final fun hexPatch (Lkotlin/jvm/functions/Function0;)Lapp/revanced/patcher/patch/RawResourcePatch;
+}
+
+public final class dropped/patches/shared/misc/hex/Replacement {
+	public static final field Companion Ldropped/patches/shared/misc/hex/Replacement$Companion;
+	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
+	public final fun replacePattern ([B)V
+}
+
+public final class dropped/patches/shared/misc/hex/Replacement$Companion {
+}
+
+public final class dropped/patches/spotify/premium/UnlockPremiumPatchKt {
+	public static final fun getUnlockPremiumPatch ()Lapp/revanced/patcher/patch/RawResourcePatch;
+}
+
+public final class dropped/patches/tasker/trial/unlock/UnlockLicensePatchKt {
+	public static final fun getUnlockLicensePatch ()Lapp/revanced/patcher/patch/BytecodePatch;
+}
+
