+++ patches/build.gradle.kts
@@ -1,0 +1,12 @@
+group = "dropped.patches"
+
+patches {
+    about {
+        name = "Dropped Patches"
+        description = "Patches for ReVanced."
+        source = "git@github.com:you/revanced-patches.git"
+        author = "indrastorms"
+        contact = "https://t.me/Indra012"
+        license = "GNU General Public License v3.0"
+    }
+}
