+++ patches/src/main/kotlin/dropped/patches/fxfileexplorer/unlockplus/unlockPlusPatch.kt
@@ -1,0 +1,24 @@
+package dropped.patches.fxfileexplorer.unlockplus
+
+import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
+import app.revanced.patcher.patch.bytecodePatch
+
+@Suppress("unused")
+val unlockPlusPatch = bytecodePatch(
+    name = "Unlock FX Plus",
+    description = "Unlock features like 'Web Access', 'Network' and 'FX Connect'.",
+) {
+    compatibleWith("nextapp.fx")
+
+    val unlockPlusPatch by isPlusUnlockedFingerprint()
+
+    execute {
+        unlockPlusPatch.mutableMethod.addInstructions(
+            0,
+            """
+            const/4 v0, 0x1
+            return v0
+        """,
+        )
+    }
+}
