+++ patches/src/main/kotlin/dropped/patches/tasker/trial/unlock/CheckLicenseFingerprint.kt
@@ -1,0 +1,7 @@
+package dropped.patches.tasker.trial.unlock
+
+import app.revanced.patcher.fingerprint
+
+internal val checkLicenseFingerprint = fingerprint {
+    strings("Can't check license")
+}
