+++ patches/src/main/kotlin/dropped/patches/tasker/trial/unlock/UnlockLicensePatch.kt
@@ -1,0 +1,17 @@
+package dropped.patches.tasker.trial.unlock
+
+import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
+import app.revanced.patcher.patch.bytecodePatch
+
+val unlockLicensePatch = bytecodePatch(
+    name = "Unlock trial",
+    description = "Unlocks the trial version.",
+) {
+    compatibleWith("net.dinglisch.android.taskerm")
+
+    val unlockLicensePatch by checkLicenseFingerprint()
+    execute {
+        unlockLicensePatch.mutableMethod.addInstruction(0, "return-void")
+        // Return the method early, which prompts the user with a non dismissible dialog, when the trial period is over.
+    }
+}
