+++ settings.gradle.kts
@@ -1,7 +1,18 @@
 rootProject.name = "dropped-patches"
 
-buildCache {
-    local {
-        isEnabled = "CI" !in System.getenv()
+pluginManagement {
+    repositories {
+        gradlePluginPortal()
+        google()
+        maven {
+            name = "GitHubPackages"
+            url = uri("https://maven.pkg.github.com/revanced/registry")
+            credentials {
+                username = providers.gradleProperty("gpr.user").getOrElse(System.getenv("GITHUB_ACTOR"))
+                password = providers.gradleProperty("gpr.key").getOrElse(System.getenv("GITHUB_TOKEN"))
+            }
+        }
     }
 }
+
+plugins { id("app.revanced.patches") version "1.0.0-dev.6" }
