--- src/main/kotlin/dropped/patches/fxfileexplorer/unlockplus/UnlockPlusPatch.kt
@@ -1,28 +1,0 @@
-package dropped.patches.fxfileexplorer.unlockplus
-
-import app.revanced.util.exception
-import app.revanced.patcher.data.BytecodeContext
-import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
-import app.revanced.patcher.patch.BytecodePatch
-import app.revanced.patcher.patch.annotation.CompatiblePackage
-import app.revanced.patcher.patch.annotation.Patch
-import dropped.patches.fxfileexplorer.unlockplus.fingerprints.IsPlusUnlockedFingerprint
-
-@Patch(
-    name = "Unlock FX Plus",
-    description = "Unlock features like 'Web Access', 'Network' and 'FX Connect'.",
-    compatiblePackages = [CompatiblePackage("nextapp.fx")]
-)
-
-@Suppress("unused")
-object UnlockPlusPatch : BytecodePatch(
-    setOf(IsPlusUnlockedFingerprint)
-) {
-    override fun execute(context: BytecodeContext) = IsPlusUnlockedFingerprint.result?.mutableMethod?.addInstructions(
-        0,
-        """
-            const/4 v0, 0x1
-            return v0
-        """
-    ) ?: throw IsPlusUnlockedFingerprint.exception
-}
