--- src/main/kotlin/dropped/patches/fxfileexplorer/unlockplus/fingerprints/IsPlusUnlockedFingerprint.kt
@@ -1,24 +1,0 @@
-package dropped.patches.fxfileexplorer.unlockplus.fingerprints
-
-import app.revanced.patcher.extensions.or
-import app.revanced.patcher.fingerprint.MethodFingerprint
-import com.android.tools.smali.dexlib2.AccessFlags
-import com.android.tools.smali.dexlib2.Opcode
-
-object IsPlusUnlockedFingerprint : MethodFingerprint(
-    returnType = "Z",
-    accessFlags = AccessFlags.PRIVATE or AccessFlags.STATIC,
-    strings = listOf(
-        "keysig= ",
-        "\nlocalsig=",
-        "nextapp.fx",
-    ),
-    opcodes = listOf(
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.MOVE_RESULT_OBJECT,
-        Opcode.CONST_16,
-        Opcode.CONST_4,
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.MOVE_RESULT_OBJECT
-    )
-)
