--- src/main/kotlin/dropped/patches/tasker/trial/unlock/fingerprints/CheckLicenseFingerprint.kt
@@ -1,7 +1,0 @@
-package dropped.patches.tasker.trial.unlock.fingerprints
-
-import app.revanced.patcher.fingerprint.MethodFingerprint
-
-object CheckLicenseFingerprint : MethodFingerprint(
-    strings = listOf("Can't check license")
-)
